refactor(patches): split main-process.js by concern behind a stable barrel
main-process.js held 22 patch functions in 1547 lines. Split into
scripts/patches/main-process/{window,tray,quit-lifecycle,browser,misc}.js;
main-process.js becomes a re-export barrel so every consumer (the patcher
CLI, descriptor files under core/, and the test suite) keeps its existing
require path and the full 22-export surface (verified equal to HEAD).

Helper clusters moved with their sole consumers: titlebar-overlay helpers
into window.js, tray-discovery + build-info helpers into tray.js,
linuxExplicitQuitExpression into quit-lifecycle.js — no cross-module
imports were needed. New files are deliberately not named patch.js, which
core/ descriptor discovery matches by exact filename.

Verified: patched output on the current upstream DMG is byte-identical to
pre-split (and to pristine main with the same feature set); 233 unit tests,
270 linux-features tests, and the full smoke suite pass.

Note (pre-existing, also reproduced on pristine main): the linux-tray patch
is not one-pass idempotent — a second patcher pass over already-patched
output rewrites the injected setLinuxTrayContextMenu electron alias
(r.Menu -> a.Menu). Harmless in the real pipeline, which always patches a
fresh DMG extraction; worth a follow-up issue.
