+++ scripts/patches/main-process/browser.js
@@ -1,0 +1,229 @@
+"use strict";
+
+const {
+  requireName,
+} = require("../shared.js");
+
+function applyBrowserUseNodeReplApprovalPatch(currentSource) {
+  const approvalPatch =
+    "startup_timeout_sec:120,tools:{js:{approval_mode:`approve`}},env:{";
+  const needle = "startup_timeout_sec:120,env:{";
+  let patchedSource = currentSource;
+  let patchedTrustedHashes = false;
+  if (patchedSource.includes(needle)) {
+    patchedSource = patchedSource.split(needle).join(approvalPatch);
+  }
+
+  const currentRuntimeConfigRegex =
+    /([A-Za-z_$][\w$]*)\.Dn\(\{([^{}]*?)nodeReplPath:([^,{}]+)(,)(?!tools:\{js:\{approval_mode:`approve`\}\})/g;
+  let patchedAnyCurrentRuntimeConfig = false;
+  patchedSource = patchedSource.replace(
+    currentRuntimeConfigRegex,
+    (_match, runtimeFactoryVar, configPrefix, nodeReplPathVar, comma) => {
+      patchedAnyCurrentRuntimeConfig = true;
+      return `${runtimeFactoryVar}.Dn({${configPrefix}nodeReplPath:${nodeReplPathVar}${comma}tools:{js:{approval_mode:\`approve\`}},`;
+    },
+  );
+
+  const trustedHashesRegex =
+    /trustedBrowserClientSha256s:([^,{}]+)\|\|([^,{}]+)\?([A-Za-z_$][\w$]*):\[\]/g;
+  patchedSource = patchedSource.replace(
+    trustedHashesRegex,
+    (match, browserUseEnabledVar, nativePipeEnabledVar, trustedHashesVar) => {
+      if (match.includes("codexLinuxTrustedBrowserClientSha256s(")) {
+        return match;
+      }
+      patchedTrustedHashes = true;
+      return `trustedBrowserClientSha256s:${browserUseEnabledVar}||${nativePipeEnabledVar}?codexLinuxTrustedBrowserClientSha256s(${trustedHashesVar}):[]`;
+    },
+  );
+
+  if (
+    patchedTrustedHashes &&
+    !patchedSource.includes("function codexLinuxTrustedBrowserClientSha256s(")
+  ) {
+    const fsVar = requireName(patchedSource, "node:fs");
+    const pathVar = requireName(patchedSource, "node:path");
+    const cryptoVar = requireName(patchedSource, "node:crypto");
+    if (fsVar == null || pathVar == null || cryptoVar == null) {
+      console.warn(
+        "WARN: Could not find fs/path/crypto aliases — skipping Linux Browser Use trusted hash patch",
+      );
+      patchedSource = patchedSource.replace(
+        /trustedBrowserClientSha256s:([^,{}]+)\|\|([^,{}]+)\?codexLinuxTrustedBrowserClientSha256s\(([A-Za-z_$][\w$]*)\):\[\]/g,
+        "trustedBrowserClientSha256s:$1||$2?$3:[]",
+      );
+      patchedTrustedHashes = false;
+    } else {
+      const helper =
+        `function codexLinuxTrustedBrowserClientSha256s(e,t=process.resourcesPath){if(process.platform!==\`linux\`)return e;let n=Array.isArray(e)?[...e]:[],r=t??"";if(r.length===0)return Array.from(new Set(n));for(let a of[\`browser\`,\`chrome\`])try{let e=(0,${pathVar}.join)(r,\`plugins\`,\`openai-bundled\`,\`plugins\`,a,\`scripts\`,\`browser-client.mjs\`);(0,${fsVar}.existsSync)(e)&&n.push((0,${cryptoVar}.createHash)(\`sha256\`).update((0,${fsVar}.readFileSync)(e)).digest(\`hex\`))}catch{}return Array.from(new Set(n))}`;
+      const strictDirective = '"use strict";';
+      const helperInsertionIndex = patchedSource.startsWith(strictDirective)
+        ? strictDirective.length
+        : 0;
+      patchedSource =
+        patchedSource.slice(0, helperInsertionIndex) +
+        helper +
+        patchedSource.slice(helperInsertionIndex);
+    }
+  }
+
+  if (
+    !patchedTrustedHashes &&
+    !patchedSource.includes("codexLinuxTrustedBrowserClientSha256s(") &&
+    patchedSource.includes("NODE_REPL_TRUSTED_BROWSER_CLIENT_SHA256S")
+  ) {
+    console.warn(
+      "WARN: Could not find Browser Use trusted hash insertion point — skipping Linux Browser Use trusted hash patch",
+    );
+  }
+
+  if (
+    patchedSource === currentSource &&
+    !patchedSource.includes(approvalPatch) &&
+    !patchedAnyCurrentRuntimeConfig &&
+    !patchedTrustedHashes &&
+    !patchedSource.includes("codexLinuxTrustedBrowserClientSha256s(")
+  ) {
+    console.warn(
+      "WARN: Could not find Browser Use node_repl config insertion point — skipping node_repl approval patch",
+    );
+  }
+
+  return patchedSource;
+}
+
+function applyLinuxBrowserUseRouteLivenessPatch(currentSource) {
+  if (currentSource.includes("codexLinuxResolveLiveBrowserUseRouteWindow")) {
+    return currentSource;
+  }
+
+  const routeWindowPattern =
+    /function ([A-Za-z_$][\w$]*)\(\{ensureWindowState:([A-Za-z_$][\w$]*),windowId:([A-Za-z_$][\w$]*),windows:([A-Za-z_$][\w$]*)\}\)\{let ([A-Za-z_$][\w$]*)=\4\.get\(\3\)\?\?null;if\(\5==null\)\{let ([A-Za-z_$][\w$]*)=([A-Za-z_$][\w$]*)\.BrowserWindow\.fromId\(\3\);\6!=null&&!\6\.isDestroyed\(\)&&!\6\.webContents\.isDestroyed\(\)&&\(\5=\2\(\6,\6\.webContents\)\)\}return \5==null\|\|\5\.window\.isDestroyed\(\)\|\|\5\.owner\.isDestroyed\(\)\?\(([A-Za-z_$][\w$]*)\(\)\.warning\(`IAB_LIFECYCLE route window is not live`,\{safe:\{hasWindowState:\5!=null,ownerDestroyed:\5\?\.owner\.isDestroyed\(\)\?\?null,windowDestroyed:\5\?\.window\.isDestroyed\(\)\?\?null,windowId:\3\},sensitive:\{\}\}\),null\):\5\}/u;
+
+  const match = currentSource.match(routeWindowPattern);
+  if (match == null) {
+    if (
+      currentSource.includes("IAB_LIFECYCLE route window is not live") &&
+      currentSource.includes("BrowserWindow.fromId")
+    ) {
+      console.warn(
+        "WARN: Could not find Browser Use route liveness helper — skipping Linux route liveness fallback patch",
+      );
+    }
+    return currentSource;
+  }
+
+  const [
+    original,
+    functionName,
+    ensureWindowStateVar,
+    windowIdVar,
+    windowsVar,
+    stateVar,
+    browserWindowVar,
+    electronVar,
+    loggerVar,
+  ] = match;
+
+  const helper = `function codexLinuxResolveLiveBrowserUseRouteWindow(e,t,n,r){if(process.platform!==\`linux\`)return null;let i=[];try{for(let e of n.values())e!=null&&!e.window.isDestroyed()&&!e.owner.isDestroyed()&&i.push(e)}catch{}if(i.length===1)return i[0];let a=[];try{a=r.BrowserWindow.getAllWindows().filter(e=>e!=null&&!e.isDestroyed()&&!e.webContents.isDestroyed())}catch{return null}if(a.length!==1)return null;let o=a[0],s=n.get(o.id)??null;return s!=null&&!s.window.isDestroyed()&&!s.owner.isDestroyed()?s:e(o,o.webContents)}`;
+  const replacement = `${helper}function ${functionName}({ensureWindowState:${ensureWindowStateVar},windowId:${windowIdVar},windows:${windowsVar}}){let ${stateVar}=${windowsVar}.get(${windowIdVar})??null;if(${stateVar}==null){let ${browserWindowVar}=${electronVar}.BrowserWindow.fromId(${windowIdVar});${browserWindowVar}!=null&&!${browserWindowVar}.isDestroyed()&&!${browserWindowVar}.webContents.isDestroyed()&&(${stateVar}=${ensureWindowStateVar}(${browserWindowVar},${browserWindowVar}.webContents))}${stateVar}==null&&(${stateVar}=codexLinuxResolveLiveBrowserUseRouteWindow(${ensureWindowStateVar},${windowIdVar},${windowsVar},${electronVar}));return ${stateVar}==null||${stateVar}.window.isDestroyed()||${stateVar}.owner.isDestroyed()?(${loggerVar}().warning(\`IAB_LIFECYCLE route window is not live\`,{safe:{hasWindowState:${stateVar}!=null,ownerDestroyed:${stateVar}?.owner.isDestroyed()??null,windowDestroyed:${stateVar}?.window.isDestroyed()??null,windowId:${windowIdVar}},sensitive:{}}),null):${stateVar}}`;
+
+  return currentSource.replace(original, replacement);
+}
+
+function applyLinuxChromeExtensionStatusPatch(currentSource) {
+  if (currentSource.includes("codexLinuxChromeProfileRoots")) {
+    return currentSource;
+  }
+
+  const fsVar = requireName(currentSource, "node:fs");
+  const osVar = requireName(currentSource, "node:os");
+  const pathVar = requireName(currentSource, "node:path");
+  if (fsVar == null || osVar == null || pathVar == null) {
+    console.warn(
+      "WARN: Could not find fs/os/path aliases — skipping Linux Chrome extension status patch",
+    );
+    return currentSource;
+  }
+
+  const unsupportedMessage =
+    "Opening Chrome extension settings is only supported on macOS and Windows";
+  const unsupportedMessageIndex = currentSource.indexOf(unsupportedMessage);
+  const openFunctionStart =
+    unsupportedMessageIndex === -1
+      ? -1
+      : currentSource.lastIndexOf("async function ", unsupportedMessageIndex);
+  const blockStart =
+    openFunctionStart === -1
+      ? -1
+      : currentSource.lastIndexOf("function ", openFunctionStart - 1);
+  const blockEnd =
+    openFunctionStart === -1
+      ? -1
+      : currentSource.indexOf("function ", openFunctionStart + "async function ".length);
+  const originalBlock = blockEnd === -1 ? null : currentSource.slice(blockStart, blockEnd);
+  if (
+    blockStart === -1 ||
+    blockEnd === -1 ||
+    !originalBlock.includes(unsupportedMessage)
+  ) {
+    console.warn(
+      "WARN: Could not find Chrome extension status functions — skipping Linux Chrome extension status patch",
+    );
+    return currentSource;
+  }
+
+  const statusFunctionName = /^function ([A-Za-z_$][\w$]*)\(\{extensionId:/.exec(
+    originalBlock,
+  )?.[1];
+  const openFunctionName = /async function ([A-Za-z_$][\w$]*)\(\{extensionId:/.exec(
+    originalBlock,
+  )?.[1];
+  const detectChromeFunctionName =
+    /detectChromeCommand:[A-Za-z_$][\w$]*=([A-Za-z_$][\w$]*)/.exec(originalBlock)?.[1];
+  const runCommandFunctionName =
+    /runCommand:[A-Za-z_$][\w$]*=([A-Za-z_$][\w$]*)/.exec(originalBlock)?.[1];
+  const extensionUrlFunctionName = /await [A-Za-z_$][\w$]*\([A-Za-z_$][\w$]*,\[([A-Za-z_$][\w$]*)\(e\)\]\)/.exec(
+    originalBlock,
+  )?.[1];
+  const macOpenFunctionName = /await [A-Za-z_$][\w$]*\(([A-Za-z_$][\w$]*),\[`-b`,/.exec(
+    originalBlock,
+  )?.[1];
+  const macBundleIdName = /await [A-Za-z_$][\w$]*\([A-Za-z_$][\w$]*,\[`-b`,([A-Za-z_$][\w$]*),/.exec(
+    originalBlock,
+  )?.[1];
+  const extensionIdValidatorName = /let [A-Za-z_$][\w$]*=([A-Za-z_$][\w$]*)\(e\),/.exec(
+    originalBlock,
+  )?.[1];
+  const profileDirFunctionName = /[A-Za-z_$][\w$]*=([A-Za-z_$][\w$]*)\(\{homeDir:/.exec(
+    originalBlock,
+  )?.[1];
+  if (
+    statusFunctionName == null ||
+    openFunctionName == null ||
+    detectChromeFunctionName == null ||
+    runCommandFunctionName == null ||
+    extensionUrlFunctionName == null ||
+    macOpenFunctionName == null ||
+    macBundleIdName == null ||
+    extensionIdValidatorName == null ||
+    profileDirFunctionName == null
+  ) {
+    console.warn(
+      "WARN: Could not identify Chrome extension status helper names — skipping Linux Chrome extension status patch",
+    );
+    return currentSource;
+  }
+
+  const replacement =
+    `function codexLinuxChromeProfileRoots({homeDir:__codexHomeDir,platform:__codexPlatform}){return __codexPlatform===\`linux\`?[(0,${pathVar}.join)(__codexHomeDir,\`.config\`,\`BraveSoftware\`,\`Brave-Browser\`),(0,${pathVar}.join)(__codexHomeDir,\`.config\`,\`google-chrome\`),(0,${pathVar}.join)(__codexHomeDir,\`.config\`,\`google-chrome-beta\`),(0,${pathVar}.join)(__codexHomeDir,\`.config\`,\`google-chrome-unstable\`),(0,${pathVar}.join)(__codexHomeDir,\`.config\`,\`chromium\`)]:[]}function codexLinuxChromeHasExtension({extensionId:__codexExtensionId,homeDir:__codexHomeDir,platform:__codexPlatform}){if(__codexPlatform!==\`linux\`)return!1;let __codexValidatedExtensionId=${extensionIdValidatorName}(__codexExtensionId);for(let __codexProfileRoot of codexLinuxChromeProfileRoots({homeDir:__codexHomeDir,platform:__codexPlatform})){if(!(0,${fsVar}.existsSync)(__codexProfileRoot))continue;for(let __codexProfileEntry of (0,${fsVar}.readdirSync)(__codexProfileRoot,{withFileTypes:!0}))if(__codexProfileEntry.isDirectory()&&(0,${fsVar}.existsSync)((0,${pathVar}.join)(__codexProfileRoot,__codexProfileEntry.name,\`Extensions\`,__codexValidatedExtensionId)))return!0}return!1}function codexLinuxChromeCommand(){let __codexPathEntries=(process.env.PATH??\`\`).split(\`:\`);for(let __codexBrowserCommand of[\`brave-browser\`,\`brave\`,\`google-chrome\`,\`google-chrome-stable\`,\`chromium-browser\`,\`chromium\`])for(let __codexPathEntry of __codexPathEntries){if(__codexPathEntry.length===0)continue;let __codexCandidate=(0,${pathVar}.join)(__codexPathEntry,__codexBrowserCommand);try{if((0,${fsVar}.existsSync)(__codexCandidate)&&(0,${fsVar}.statSync)(__codexCandidate).isFile())return __codexCandidate}catch{}}return null}function ${statusFunctionName}({extensionId:__codexExtensionId,homeDir:__codexHomeDir=(0,${osVar}.homedir)(),localAppDataDir:__codexLocalAppDataDir=process.env.LOCALAPPDATA,platform:__codexPlatform=process.platform}){if(__codexPlatform===\`linux\`)return codexLinuxChromeHasExtension({extensionId:__codexExtensionId,homeDir:__codexHomeDir,platform:__codexPlatform});let __codexValidatedExtensionId=${extensionIdValidatorName}(__codexExtensionId),__codexProfileDir=${profileDirFunctionName}({homeDir:__codexHomeDir,localAppDataDir:__codexLocalAppDataDir,platform:__codexPlatform});return __codexProfileDir==null||!(0,${fsVar}.existsSync)(__codexProfileDir)?!1:(0,${fsVar}.readdirSync)(__codexProfileDir,{withFileTypes:!0}).some(__codexProfileEntry=>__codexProfileEntry.isDirectory()&&(0,${fsVar}.existsSync)((0,${pathVar}.join)(__codexProfileDir,__codexProfileEntry.name,\`Extensions\`,__codexValidatedExtensionId)))}async function ${openFunctionName}({extensionId:__codexExtensionId,platform:__codexPlatform=process.platform,detectChromeCommand:__codexDetectChromeCommand=${detectChromeFunctionName},runCommand:__codexRunCommand=${runCommandFunctionName}}){if(__codexPlatform===\`darwin\`){await __codexRunCommand(${macOpenFunctionName},[\`-b\`,${macBundleIdName},${extensionUrlFunctionName}(__codexExtensionId)]);return}if(__codexPlatform===\`win32\`){let __codexChromeCommand=__codexDetectChromeCommand();if(__codexChromeCommand==null)throw Error(\`Google Chrome is not installed\`);await __codexRunCommand(__codexChromeCommand,[${extensionUrlFunctionName}(__codexExtensionId)]);return}if(__codexPlatform===\`linux\`){let __codexChromeCommand=codexLinuxChromeCommand()??__codexDetectChromeCommand();if(__codexChromeCommand==null)throw Error(\`Google Chrome, Brave, or Chromium is not installed\`);await __codexRunCommand(__codexChromeCommand,[${extensionUrlFunctionName}(__codexExtensionId)]);return}throw Error(\`Opening Chrome extension settings is only supported on macOS, Windows, and Linux\`)}`;
+
+  return currentSource.slice(0, blockStart) + replacement + currentSource.slice(blockEnd);
+}
+
+module.exports = {
+  applyBrowserUseNodeReplApprovalPatch,
+  applyLinuxBrowserUseRouteLivenessPatch,
+  applyLinuxChromeExtensionStatusPatch,
+};
